requirements.txt
numpy
pandas
scikit-learn
pytest
requests
fastapi==0.65.2
uvicorn
gunicorn
tensorflow-cpu
pyyaml
dvc
dvc[s3]
